+++ CHANGELOG.md
@@ -4,6 +4,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 # Upcoming
 
 ### 🔄 Changed
+- Reactions popup disabled if channel is frozen
 
 # [4.35.1](https://github.com/GetStream/stream-chat-swiftui/releases/tag/4.35.1)
 _August 10, 2023_